refactor(LottoPaperTest, LottoTest): 패키지 이동
